Exercice 4.15-4.23

+++ part4/liste_de_blog/controllers/bloges.js
@@ -2,11 +2,17 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/bloges')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   if (!request.body.title || !request.body.url) {
     return response.status(400).json({ error: 'title and url are required' })
   }
@@ -15,14 +21,34 @@ blogsRouter.post('/', async (request, response) => {
     title: request.body.title,
     author: request.body.author,
     url: request.body.url,
-    likes: request.body.likes
+    likes: request.body.likes,
+    user: user._id
   })
 
   const saved = await blog.save()
+  user.blogs = user.blogs.concat(saved._id)
+  await user.save()
+
   response.status(201).json(saved)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'only the creator can delete this blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
